requirements.txt
flask>=2.3.3
requests>=2.31.0
gunicorn>=21.2.0
gevent>=22.10.2
google-cloud-storage>=2.14.0
google-genai
python-dotenv>=1.0.0